requirements.txt
flask
requests
python-dotenv
gunicorn
stripe
deepface
scipy
numpy
opencv-python
pillow
tensorflow>=2.5.0
tf-keras
psycopg2-binary